requirements.txt
asgiref==3.9.0
Django==4.2.2
django-filter==25.1
django-phonenumber-field==8.1.0
django-phonenumbers==1.0.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dotenv==0.9.9
phonenumber-field==0.0.1
phonenumbers==9.0.8
pillow==11.3.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dotenv==1.1.1
sqlparse==0.5.3
tzdata==2025.2